Support uint32 array buffers

+++ source/web/xatlas_web.cpp
@@ -39,8 +39,8 @@ MeshBufferInfo createMesh(uint32_t vertexCount, uint32_t indexCount, bool normal
     meshDecl->vertexCount = vertexCount;
     meshDecl->indexCount = indexCount;
 
-    meshDecl->indexData = new uint16_t[indexCount];
-    meshDecl->indexFormat = xatlas::IndexFormat::UInt16;
+    meshDecl->indexData = new uint32_t[indexCount];
+    meshDecl->indexFormat = xatlas::IndexFormat::UInt32;
     meshBufferInfo.indexOffset = (uint32_t) meshDecl->indexData;
 
     meshDecl->vertexPositionData = new float[vertexCount * 3];
